requirements.txt
asgiref==3.3.4
Django==3.2.3
djangorestframework==3.12.4
grpcio==1.38.0
protobuf==3.17.1
pytz==2021.1
six==1.16.0
uWSGI==2.0.18
django-cors-headers==3.7.0
python-decouple==3.4
psycopg2-binary==2.8.6
pytz==2020.5
sqlparse==0.4.1

